feat: 🎸 register_user で user を返す

+++ kernel/src/model/user.rs
@@ -8,7 +8,7 @@ pub struct User {
     pub traq_id: String,
     pub traq_display_id: String,
     pub channel_id: String,
-    pub balance: i64,
+    pub balance: i32,
 }
 
 #[derive(new, Debug)]
